feat - entity 노출 대신 DTO 변경      - Student에 관한 CRUD 오나료

+++ src/main/java/com/student/domain/Student.java
@@ -13,15 +13,13 @@ import java.util.List;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
-@EqualsAndHashCode(of = "id")
-@Entity
-@Setter
+@Getter @EqualsAndHashCode(of = "id")
+@Entity @Setter
 public class Student {
     @Id
     @GeneratedValue
     @Column(name = "student_id")
-    private Integer id;
+    private Long id;
 
     private String name;
     private int age;
